+++ plugin_hs/run.go
@@ -17,8 +17,7 @@ import (
 	"github.com/FloatTech/ZeroBot-Plugin/utils/file"
 )
 
-var botpath, _ = os.Getwd()
-var cachedir = botpath + "/data/hs/"
+var cachedir = file.BOT_PATH + "/data/hs/"
 
 var header = req.Header{
 	"user-agent": `Mozilla/5.0 (Linux; Android 6.0; Nexus 5 Build/MRA58N) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/86.0.4240.198 Mobile Safari/537.36`,
