+++ utils/file/dir_nowin.go
@@ -1,0 +1,11 @@
+//go:build !windows
+// +build !windows
+
+package file
+
+import "os"
+
+func Pwd() (path string) {
+	path, _ = os.Getwd()
+	return
+}
