+++ utils/file/dir_win.go
@@ -1,0 +1,14 @@
+//go:build windows
+// +build windows
+
+package file
+
+import (
+	"os"
+	"strings"
+)
+
+func Pwd() string {
+	path, _ := os.Getwd()
+	return strings.ReplaceAll(path, "\\", "/")
+}
