+++ utils/file/dl.go
@@ -21,15 +21,3 @@ func DownloadTo(url, file string) error {
 	}
 	return err
 }
-
-// IsExist 文件/路径存在
-func IsExist(path string) bool {
-	_, err := os.Stat(path)
-	return err == nil || os.IsExist(err)
-}
-
-// IsExist 文件/路径不存在
-func IsNotExist(path string) bool {
-	_, err := os.Stat(path)
-	return err != nil && os.IsNotExist(err)
-}
